+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Product Spec is generated based on the `skuSpecs` from each product.
 - Add `productStatus` as `active` always.
 - Add `skuIsActive` as `true` always.
+- Add `skuName` generated from `productName` + values of the `skuSpecs`
 
 ### Fixed
 
@@ -25,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `productSlug` column from csv and generate it based on productName.
 - Remove `productSpecs (name_1 to name_5 and values_1 to values_5)` columns from csv and generate it based on `skuSpecs`.
 - Remove `productStatus, skuIsActive` and add value as default.
+- Remove `skuName`
 
 ## [1.2.1] - 2024-04-03